test: for gateway, head and worker runtime grpc args
